# openapi.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rhythmeta backend API",
    "description": "Auto-generated API specification from registered Hono routes and TypeScript route-source analysis. Schemas are inferred from actual handler input/output types.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://api.rhythmeta.org",
      "description": "Current backend server"
    }
  ],
...
  "paths": {
...
    "/maimaid/v1/backups": {
      "get": {
        "operationId": "get_maimaid_v1_backups",
        "tags": [
          "maimaid"
        ],
        "summary": "GET /maimaid/v1/backups",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {}
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post_maimaid_v1_backups",
        "tags": [
          "maimaid"
        ],
        "summary": "POST /maimaid/v1/backups",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {}
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/maimaid/v1/backups/{id}/commit": {
      "post": {
        "operationId": "post_maimaid_v1_backups_id_commit",
        "tags": [
          "maimaid"
        ],
        "summary": "POST /maimaid/v1/backups/{id}/commit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {}
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/maimaid/v1/backups/{id}": {
      "delete": {
        "operationId": "delete_maimaid_v1_backups_id",
        "tags": [
          "maimaid"
        ],
        "summary": "DELETE /maimaid/v1/backups/{id}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {}
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/chunithmd/v1/backups": {
      "get": {
        "operationId": "get_chunithmd_v1_backups",
        "tags": [
          "chunithmd"
        ],
        "summary": "GET /chunithmd/v1/backups",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {}
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post_chunithmd_v1_backups",
        "tags": [
          "chunithmd"
        ],
        "summary": "POST /chunithmd/v1/backups",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {}
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/chunithmd/v1/backups/{id}/commit": {
      "post": {
        "operationId": "post_chunithmd_v1_backups_id_commit",
        "tags": [
          "chunithmd"
        ],
        "summary": "POST /chunithmd/v1/backups/{id}/commit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {}
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/chunithmd/v1/backups/{id}": {
      "delete": {
        "operationId": "delete_chunithmd_v1_backups_id",
        "tags": [
          "chunithmd"
        ],
        "summary": "DELETE /chunithmd/v1/backups/{id}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {}
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}